+++ packages/renderer-webgl/src/glsl.d.ts
@@ -7,3 +7,16 @@ declare module "*.vert" {
   const source: string;
   export default source;
 }
+
+declare module "*?royal-codec-url" {
+  const url: string;
+  export default url;
+}
+
+interface ImportMetaEnv {
+  readonly SSR: boolean;
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv;
+}